fix: 엔티티 값 교체

+++ src/main/java/com/kickmate/kickmate/domain/commentary/entity/RawActionEvent.java
@@ -17,7 +17,7 @@ public class RawActionEvent {
 
     @NotNull
     @Column(name = "game_id", nullable = false)
-    private Integer gameId;
+    private Long gameId;
 
     @NotNull
     @Column(name = "action_id", nullable = false)
